# docs/swagger.yaml
definitions:
  controller.response:
    properties:
      message:
        type: string
    type: object
...
  models.Admins:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      lastVisit:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  models.News:
    properties:
      created:
        $ref: '#/definitions/models.Admins'
      createdAt:
        type: string
      descriptionEn:
        type: string
      descriptionRu:
        type: string
      descriptionUz:
        type: string
      id:
        type: integer
      image:
        type: string
      nameEn:
        type: string
      nameRu:
        type: string
      nameUz:
        type: string
      position:
        type: integer
      updated:
        $ref: '#/definitions/models.Admins'
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server CRM server.
  title: Golang CRM Swagger Documentation
  version: "1.0"
paths:
...
  /api/news:
    get:
      consumes:
      - application/json
      description: this api is get news
      parameters:
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.News'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Get news
      tags:
      - News
    post:
      consumes:
      - application/json
      description: this api is create news
      parameters:
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: position
        type: integer
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create news
      tags:
      - News
  /api/news/{id}:
    delete:
      consumes:
      - application/json
      description: this api is delete news
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: DELETE news
      tags:
      - News
    get:
      consumes:
      - application/json
      description: this api is Update news
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Update news
      tags:
      - News
    put:
      consumes:
      - application/json
      description: this api is Update news
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: string
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: position
        type: integer
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Update news
      tags:
      - News
...
securityDefinitions:
  BearerAuth:
    description: AUTH.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"